clk: qcom: clk-regmap: Update type for rate in clk prepare
On 32 bit the PLL frequency can be beyond the range
of the `int` leads to improper assignment, so update
rate type to 'unsigned long'.

+++ drivers/clk/qcom/clk-regmap.c
@@ -195,7 +195,7 @@ EXPORT_SYMBOL(clk_post_change_regmap);
 int clk_prepare_regmap(struct clk_hw *hw)
 {
 	struct clk_regmap *rclk = to_clk_regmap(hw);
-	int rate = clk_hw_get_rate(hw);
+	unsigned long rate = clk_hw_get_rate(hw);
 	int vdd_level;
 
 	if (!rclk->vdd_data.rate_max)
